sinowealth7: Skeleton driver for "Sinowealth 7"
This driver detects my "Easterntimes Tech" T7 gaming mouse which appears
on USB with manufacturer "SINOWEALTH", product "Game Mouse", VID 258a
and PID 1007. It contains various bits copied from the sinowealth
driver, but the two appear not to be compatible.

The driver only supports turning the LED on and off currently, but it's
a start. (The same nibble can also be used to control the brightness.)